+++ src/snowflake/cli/plugins/snowpark/package/commands.py
@@ -22,7 +22,7 @@ from snowflake.cli.plugins.snowpark.snowpark_shared import PackageNativeLibrarie
 
 app = SnowTyper(
     name="package",
-    help="Manage custom Python packages for Snowpark",
+    help="Manages custom Python packages for Snowpark",
 )
 log = logging.getLogger(__name__)
 
@@ -85,7 +85,7 @@ def package_upload(
         False,
         "--overwrite",
         "-o",
-        help="Whether to overwrite the file if it already exists.",
+        help="Overwrites the file if it already exists.",
     ),
     **options,
 ) -> CommandResult: