+++ tests/app/test_telemetry.py
@@ -10,8 +10,9 @@ from snowflake.connector.version import VERSION as DRIVER_VERSION
 @mock.patch("snowflake.cli.app.telemetry.platform.platform")
 @mock.patch("snowflake.cli.app.telemetry.get_time_millis")
 @mock.patch("snowflake.connector.connect")
+@mock.patch("snowflake.cli.plugins.connection.commands.ObjectManager")
 def test_executing_command_sends_telemetry_data(
-    mock_conn, mock_time, mock_platform, mock_version, runner
+    _, mock_conn, mock_time, mock_platform, mock_version, runner
 ):
     mock_time.return_value = "123"
     mock_platform.return_value = "FancyOS"